rendering user's photos, fixing moduls
